attempt to separate prod & dev code
Based on the hostname & inventory variables, either ship a prod
or a dev site (or really, just ship custom env for whatever the
site is, not just prod & dev).

Also, actually use the letsencrypt_email variable that was
already in the vars file but not in use by the code/templates.